+++ debug/debug.py
@@ -2,17 +2,16 @@
 from __future__ import unicode_literals, print_function
 
 import argparse
-import io
 import json
-import os
 from builtins import input, bytes
+from pathlib import Path
 
 from snips_nlu import SnipsNLUEngine, load_resources
 from snips_nlu.pipeline.configs.nlu_engine import NLUEngineConfig
 
 
 def debug_training(dataset_path, config_path=None):
-    with io.open(os.path.abspath(dataset_path), "r", encoding="utf8") as f:
+    with Path(dataset_path).open("r", encoding="utf8") as f:
         dataset = json.load(f)
 
     load_resources(dataset["language"])
@@ -20,7 +19,7 @@ def debug_training(dataset_path, config_path=None):
     if config_path is None:
         config = NLUEngineConfig()
     else:
-        with io.open(config_path, "r", encoding="utf8") as f:
+        with Path(config_path).open("r", encoding="utf8") as f:
             config = NLUEngineConfig.from_dict(json.load(f))
 
     engine = SnipsNLUEngine(config).fit(dataset)
@@ -35,7 +34,7 @@ def debug_training(dataset_path, config_path=None):
 
 
 def debug_inference(engine_path):
-    with io.open(os.path.abspath(engine_path), "r", encoding="utf8") as f:
+    with Path(engine_path).open("r", encoding="utf8") as f:
         engine_dict = json.load(f)
 
     load_resources(engine_dict["dataset_metadata"]["language_code"])