+++ snips_nlu/cli/dataset/__init__.py
@@ -1,0 +1,1 @@
+from snips_nlu.cli.dataset.assistant_dataset import AssistantDataset