+++ snips_nlu_dataset/__init__.py
@@ -1,2 +1,0 @@
-from snips_nlu_dataset.assistant_dataset import (
-    AssistantDataset, main_generate_dataset)