+++ snips_nlu_samples/sample.py
@@ -1,15 +1,14 @@
 from __future__ import unicode_literals, print_function
 
-import io
 import json
-from os.path import dirname, abspath, join
+from pathlib import Path
 
 from snips_nlu import SnipsNLUEngine, load_resources
 from snips_nlu.default_configs import CONFIG_EN
 
-SAMPLE_DATASET_PATH = join(dirname(abspath(__file__)), "sample_dataset.json")
+SAMPLE_DATASET_PATH = Path(__file__).parent / "sample_dataset.json"
 
-with io.open(SAMPLE_DATASET_PATH) as f:
+with SAMPLE_DATASET_PATH.open() as f:
     sample_dataset = json.load(f)
 
 load_resources("en")